feat: gate snapshot-bootstrap surface behind `unsafe-dump-load` feature
Three escape-hatch APIs added for caller-driven subtree dump/restore are
now Cargo-feature-gated so production builds don't carry any of them:

  * `GroveDb::raw_storage()`               — borrow underlying RocksDbStorage
  * `RocksDbStorage::ingest_subtree_sst()` — bulk-load SST into a CF
  * `GroveDb::replace_subtree_root()`      — caller-provided child hash

Also generalizes the third API. The previous
`replace_commitment_tree_subtree_root` took CommitmentTree-specific
arguments (total_count, chunk_power, flags) and constructed the Element
internally. The same Merk-tail plumbing is identical across all non-Merk
tree types (commitment / mmr / bulk-append / dense), so the new shape
just takes an arbitrary `Element` plus the caller-computed combined root:

  pub fn replace_subtree_root<'b, B, P>(
      &self,
      path: P,
      key: &[u8],
      new_element: Element,
      new_combined_root: [u8; 32],
      transaction: TransactionArg,
      grove_version: &GroveVersion,
  ) -> CostResult<(), Error>

The body cheap-checks that the Element is a tree variant (rejecting
Item/Reference, which have no child-hash slot) but leaves
hash-vs-state correctness to the caller — same contract as before,
just no longer commitment-tree-specific.
