+++ grovedb/src/operations/mod.rs
@@ -26,5 +26,11 @@ pub mod bulk_append_tree;
 #[cfg(feature = "minimal")]
 pub mod dense_tree;
 
+/// Caller-driven subtree-root replacement. Bypasses grovedb's normal
+/// "compute child hash from subtree state" invariant — see the module-level
+/// docs in `replace_subtree_root.rs` for the safety contract.
+#[cfg(all(feature = "minimal", feature = "unsafe-dump-load"))]
+pub mod replace_subtree_root;
+
 #[cfg(feature = "minimal")]
 pub use get::{QueryItemOrSumReturnType, MAX_REFERENCE_HOPS};
